Use @fastify/jwt instead of jsonwebtoken with custom code
