+++ src/server.ts
@@ -1,11 +1,10 @@
 import Fastify from 'fastify';
 
 import { serializerCompiler, validatorCompiler } from 'fastify-type-provider-zod';
-import { JWT_SECRET, NODE_ENV } from './config';
+import { NODE_ENV } from './config';
 import { AppDataSource } from './db';
 import loggingByEnv from './logging';
-import parseJwtToken from './plugins/parse-jwt-token';
-import requireJwtToken from './plugins/require-jwt-token';
+import setupJwtTokenAuth from './plugins/authenticate';
 import apiV1 from './routes';
 
 export async function createServer() {
@@ -19,11 +18,7 @@ export async function createServer() {
   fastify.setValidatorCompiler(validatorCompiler);
   fastify.setSerializerCompiler(serializerCompiler);
 
-  fastify.register(parseJwtToken, {
-    algorithms: ['HS256'],
-    secret: JWT_SECRET,
-  });
-  fastify.register(requireJwtToken);
+  fastify.register(setupJwtTokenAuth);
 
   fastify.register(apiV1);
