+++ src/services/AuthService.ts
@@ -1,5 +1,3 @@
-import jwt from 'jsonwebtoken';
-import { JWT_SECRET } from '../config';
 import { getRepository } from '../db';
 import { User } from '../models/User';
 import { Token } from '../schemas/Auth';
@@ -7,7 +5,7 @@ import { AppError, AppErrorCode } from '../utils/ApplicationError';
 import { passwordEncryptionService } from './PasswordEncryptionService';
 
 class AuthService {
-  async signIn(email: string, plainPassword: string) {
+  async signIn(email: string, plainPassword: string): Promise<Token> {
     const dbUser = await getRepository(User).findOne({
       where: { email },
     });
@@ -18,17 +16,11 @@ class AuthService {
     if (!isPasswordCorrect) {
       throw new AppError(AppErrorCode.WRONG_EMAIL_OR_PASSWORD, 'User with email/password pair not found');
     }
-    return jwt.sign(
-      {
-        sub: dbUser.id,
-        role: dbUser.role,
-        exp: Math.floor(Date.now() / 1000) + 60 * 60 * 24 * 7 /* 1 week */,
-      } satisfies Token,
-      JWT_SECRET,
-      {
-        algorithm: 'HS256',
-      },
-    );
+    return {
+      sub: dbUser.id,
+      role: dbUser.role,
+      exp: Math.floor(Date.now() / 1000) + 60 * 60 * 24 * 7 /* 1 week */,
+    };
   }
 }
